[DELETE] RoomInfoView View안으로 (#514)
